requirements.txt
Pillow==10.2.0
requests_toolbelt==1.0.0
websocket_client==1.7.0
firebase-admin
